macOS: change quit dialog to handle return key
Ghostty does handle `Cmd+q` to quit the application on macOS, but the
dialog it presented didn't accept the return key to confirm.

That drove me a bit mad, so I went down the rabbithole to find out how
to handle that correctly.

Turns out: `.keyboardShortcut(.defaultAction)` _is_ the recommended way,
but it doesn't work on macOS. At least not on macOS 13. I also found
comments saying that it stopped working on macOS 12.

So, what this here does it to replace the SwiftUI `confirmationDialog`
with an `NSAlert` that pops up whenever the user presses `Cmd+q`.

It's up to you to accept this or not. I'd understand if you want to keep
this SwiftUI only.

My motiviation is it did annoy me quite a bit having to reach for the
mouse when hitting `Cmd+q` :)

Drawback: The dialog doesn't show up in the middle of the window, it
shows up top-middle of the screen. Actually, I'm not sure how NSAlert
position is determined.