macos: associate SurfaceViews with UUIDs
Add a unique ID (UUID) to each surface view and use it to associate
SurfaceViews with notifications. This is safer than using the address of
the surface view as the identifier, as there is no risk of addressing
invalid memory in the case where the view is freed while the
notification is still active.