Surface: clear selection on all key input, not just text-generated
This removes the check to see if a keypress generated any UTF8 data
before clearing the selection.

This means that *any* keypress that is a non-modifier, sent to the
terminal, now clears the selection, which should resolve a number of
situations where selection was not being properly cleared (e.g. vim
with set mouse="", less, and shell history navigation).