termio: Explicitly set SHELL env var
In my experience, and maybe I'm wrong, the terminal sets SHELL on the
process it starts up so the shell is aware of what it is.

That allows predictable use when doing something maybe like: `exec
$SHELL` to "reload current shell".

I found this wasn't being set when I override `command` in the config. I
would, and have expected, SHELL to be set to match that value.

For example, I set `command = /opt/homebrew/bin/zsh` and expected that
to be present as my SHELL.