Move libadwaita to an app runtime
libadwaita is a very opinionated library, hence why it was an optional
build option before. Previously however, ghostty was only using
libadwaita as a way to theme the GTK app runtime. libadwaita is much
more than a theme for GTK. It is also library full of widgets. For
instance, GtkNotebook is an inferior tab management widget to
AdwTabView, which I plan to enable in the future for the libadwaita
runtime.

In a further off future, we can also depend on libpanel,
a libadwaita-based GTK library for managing panels, in order to provide
our splits. The GTK app runtime would have to come up with a different
solution, most likely involving GtkGridView or some such.