Fix spurious warnings introduced in #1182 if the command has arguments.
Since #1102 the "command" configuration can contain more than just an
executable - it can contain arguments to the command. The changes in #1182
treat the whole command (including any arguments) as the name
of the executable. This will cause spurious configuration warnings if a
command with arguments is specified even if the actual executable exists
on the PATH or on the filesystem.

This PR corrects this by splitting the command at the first space
character and checking for the first part of the command in the PATH or
on the filesystem. This is a bit of a "hack". The actual command will
eventually be parsed by the shell so shell quoting may make Ghostty warn
about valid shell commands.

This PR also splits the command checking out into a separate function
to make testing easier. Several tests for the correct function of the
command finalization are added as well.