parser: allow semicolons in any sequence
The parser was gated to only allow semicolons in 'm' (SGR) sequences.
This prevents an input parser from receiving kitty encoded key events.
Remove the check for semicolons. This should have no side effects: there
are only a few sequences which use ':' delimiters. If an application
were to send one that Ghostty doesn't *already* know about, we would
likely fail to handle the sequence in the handler portion.