terminal(parser): refactor to enable subparameter tracking
Refactor the vt parser to track each param as either a subparameter or
not. This is done by keeping a bitset alongside the params array. This
bitset  is a direct map of param to subparam state. This enables easy
checking in the SGR parser for subparams.

Modify SGR parser to match the new csi_dispatch struct. This actually
worked to simplify the handling, and should make it more robust. For
example, the following SGR sequence is now able to be parsed by ghostty
(previously, it would fail): `\x1b[4:2;1m`.