D-Bus and SystemD activation
This updates Ghostty to use D-Bus and SystemD activation to start.
In addition to making future Gnome/GTK integration easier, this also
implements a "background daemon" mode (Fixing #2010). The configuration
variable `quit-after-last-window-closed` is changed to an enum that
takes `always`, `never`, and `after-timeout`. Always works like
`true` used to and `never` works like `false` used to. If set to
`after-timeout` Ghostty will quit a configurable delay after the last
surface is closed. This is set by default to five minutes. The timeout
only works on Linux for now.

I attempted to update the macOS code to support the change to
`quit-after-last-window-closed` but I have no way to test the changes so
I didn't attempt to implement the background timeout.