renderer: decouple text blink from cursor blink
Coupling the text blink to the cursor blink leads to interesting
problems, such as the text blink stopping when the user is typing,
as the cursor is coded to stop blinking during typing. We now make the
text blink on a separate timer that drives the cursor blink, so that
text input events can easily cancel the cursor blink, waiting for it to
be re-synchronized to the text blink.