gtk: add "search with selection" keybind action and context menu item
This adds the ability to launch an internet search with any selected
text on the active surface. If there is no selected text, the action
is a no-op. Search providers are configured using the `search-provider`
configuration entry and can be chosen from a set of predefined options
or a supplied custom URI.

This only implements the context menu for GTK but extending this to
macOS should be fairly easy as most of the logic is in the core. The
keybind action should work on macOS as-is although it is not assigned a
keybind by default.