Fix copying the theme name
Prior to this change both `C` and `c` would copy the path to the theme even
though the help screen claimed that `c` would copy the theme name.

The problem was that the codepoint for `C` is `C` not `c`. The fact that the
shift modifier is coming along is not sufficient to distinguish the two cases.

Tested:
  Before the change: 'c' copies path and 'C' copies path
	After the change: 'c' copies the name and 'C' copies the path