macos: Add "Use Option as Meta Key" menu item
Fixes #2520.

Makes state per surface, to match Terminal.app

Adds menu item to xib and wires it up.

Clicking the menu item sends an action into libghostty,
which toggles state on the surface and sends a notification
back. The menu item state then updates in response to that
notification.

The menu item also updates when the current surface changes,
either if the current split changes, or when tabs change, or
when the quick terminal is opened or closed.