reorganize logic to make better semantic sense
This should be functionally identical, it just makes more sense written
this way. When releasing the style we simultaneously clear the ID back
to default, and after that point it's guaranteed to be default until
changed again so we don't need to set it before returning if the new
style is default or before adding the style to the new page.