testing: update tests to point a ~real binary
`/usr/bin/env` does not exist in a nix build sandbox. This only works as
a test binary when running in `nix develop` as it shares its environment
with the user

`/bin/sh` isn't amazing as a target as it's relying on stdenv creating
it in a build sandbox and users having while running `nix develop`.
Alternatives require build flags or some sort of contract with
packaging.