Use foreground color in macOS tab titles
Currently tab titles will use the default macOS tab foreground color,
but the color schemes background color. This can result in tabs with
unreadable text.

This change updates the macOS UI tabs to use the foreground color
defined in the config, or the overridden value when set dynamically.

This change:

- Tracks `foregroundColor` in the SurfaceView which is pre-populated
  with the config foreground color along with support for color change
  update events.
- Adds a new `titleForegroundColor` property to use in `TerminalWindow`
  titles.
- Updates each tab window foreground color when `syncAppearance` is
  called.
- Calls `syncAppearance` when a window becomes the key window.