cli/list_themes: Sort themes by background color
This sorts theme background colors into 16 buckets, from dark to light
in ascending order (with the name as the final tie-breaker as before).
It's intended to prevent the jarring (epilepsy-inducing even?) flashes
that happen when you hold down the down/up arrow in the list-themes UI.
It's still not perfect since the foreground colors can vary wildly
between themes, but it should be significantly better than the
white-to-black (and vice-versa) flashes before this change.

Because we need to determine the theme's background color when sorting,
this change moves the loading of the theme files to the initialization
step instead of doing it just-in-time. To preserve the message box
displayed when the theme cannot be read, we store the error message as
part of the ThemeListElement.