ansi: Add support blinking text
Add support for the blinking text ansi character. Both "slow" (code 5)
and "fast" (code 6) blink at the same interval as the cursor, as it
seems implementations often do this.

Blinking of text stops when the screen loses focus, same as the cursor.
To keep blinking of text and cursor in sync, typing a character no
longer resets the cursor blinking timer. Instead of reseting the timer,
the cursor is immediately set to visible and will stay visible for at
least the length of one interval. This has the side-effect that at
startup, it takes two intervals until the first cursor blink happens.