bash: prefer $HOME over tilde notation
This reverts part of #5075. While the ~ notation is more concise, I
think the consensus is that using $HOME is more explicit, so prefer it.

We've also seen reports that something about #5075 breaks the startup
sequence in some environments. Reverting this part of the change will
help us narrow that down.