sextants: always properly scale the middle
what's going on here is we're saying if there's a
single extra row in the cell to be distributed,
put it on the center, and if there are two, put them
on the top and bottom (we get the bottom without
specification because we use the cell geometry
directly there). in the mod 2 case, we still want
to draw the middle one line past its natural position,
to account for the extra line from above. otherwise
your bottom third is one larger than your top, and
two larger than your middle, which no one wants.