feat: add support for relative paths
to make relative path clickable we'll check whether we're able to
resolve path in local system or not? We check line and it's all
substring seperated by <space>. If path is resolvable then we will build
the selection and return the result i.e. `ActionSelection` struct. `ActionSelection`
is contains same `action` and `selection` properties.