bash: use a shell command for shell integration
Prior to #7044, on macOS, on macOS, our shell-integrated command line
would be executed under `exec -l`, which causes bash to be started as a
login shell. This matches the macOS platform norms.

The change to direct command execution meant that we'd skip that path,
and bash would start as a normal interactive (non-login) shell on macOS.

Restore the previous behavior by returning a shell command from our bash
integration routine. This is consistent with the code path used when
shell integration is disabled.

We could alternatively add `--login` to our shell-integrated bash
command line for the macOS build. That would let us continue using the
direct command strategy, but it would also result in having coordinated,
conditional behavior in two places (shell integration? macOS?).