termio: exec direct commands as login sessions
When executing a .direct command under macOS, we pass the arguments
directly to login. We want this to be a login session so we don't want
to use login's `-l` flag (which suppress the login session behavior).

We continue to use `login -l` for .shell commands because we wrap them
in a call to `exec -l`, which handles the login session itself.