PageList.zig: Allow a Pin's x to be a column or negative.
Changes Pin.x from u16 to a union that's either .col: u16 (common
case) or .neg: void (newly introduced case). The .neg situation occurs
when the mouse is dragging and it's at the very left edge of the
terminal, or even further left.

This also changes gtkMouseMotion to not clamp x motion to 0 or
positive. This is to ensure that a slightly positive x value that
transitions negative will trigger `!is_cursor_still`.

Also, this adds a couple unit tests.

The overall goal of the change is to fix this issue where an attempt
to select a full line of text by dragging the mouse off the left hand
side would inadvertantly select one additional character:

https://github.com/ghostty-org/ghostty/discussions/5058