BOO! Ids. (Booids, rhymes with "squids")
This introduces an ID format for Ghostty objects, lovingly called the
"booid" (rhyming with "squids" because "guids"). I'm not a huge fan of
overly cute naming -- I prefer obvious names -- but all the obvious
names like "unique ID" or "global ID" are too similar to other, very
real, standard ID formats.

These IDs will be used in the short term by app IDs to reference
surfaces for things like notifications. And in the long term, we'll use
these for the still-to-be-designed Ghostty API. And super long term,
we'll use these for session attach/detach.

There's no concurrency for our ID generation today, so this commit only
has a simple non-thread-safe implementation. That's all we need.

Ghostty's use case has slightly different requirements than most unique
ID requirements, so I believe a custom format is the way to go here.
These requirements are noted here:

The booid is a 64-bit ID to make it easy to send across C ABIs and local
transports like D-Bus, both of which are very important for a desktop
application like Ghostty.

We follow the same timestamp, machine, sequence number format as
Snowflake IDs, but we use the full 64 bits for the ID, rather than 63,
since can explicitly note its an unsigned ID format. Our bit width
breakdown allows for 1,024 nodes in a cluter producing 4,096 IDs per
millisecond.

"A cluster?! What the @#!* are you talking about? This is a local
terminal emulator." Yes, yes indeed. 42-bits of timestamp and 12 bits
of sequence are more than enough for a local terminal emulator, so we
have 10 bits left over (arguably, we need even less sequence bits). For
now, we designate this as a "machine ID" and always set it to "0" since
we are local-only.

But I think this will be useful in the future if/when we introduce the
ability to attach to remote Ghostty instances for session sharing,
restore, etc. This is a use case and feature I have always planned. 

The question of how we assign machine IDs in that future is left to the
future. I propose a simple idea in the comments in this commit: assign
and map them locally! All local IDs can produce machine ID 0 and then
the CLIENT side can replace those bits with some value. This eliminates
a distributed consensus problem and allows a client terminal to attach
to up to 1,024 remote instances at once (unlikely lol).