build: build ReleaseFast and baseline CPU by default for system packages
This makes `--system` imply `-Doptimize=ReleaseFast` and
`-Dcpu=baseline`. These can still be overridden by the user by setting
the same flags explicitly.

The `--system` flag is used to build system packages and its a safe
assumption that in those cases, Ghostty is being built for general use
amongst many users, so the default should be optimized for
performance and compatibility.

This was always documented in our PACKAGING.md, but is now automatic.
The PACKAGING.md file has been updated to reflect this.