gtk-ng: deprecate gtk-single-instance=desktop
This fixes the problem of running a bare `ghostty` at the CLI (or from
any of the numerous launchers used by tiling window managers) creating
a new instance of Ghostty rather than sending a signal to an existing
Ghostty instance to open up a new window.

This happened because when `gtk-single-instance=desktop` and you ran
`ghostty` from the CLI, it was treated as `gtk-single-instance=false`
which would create a new Ghostty instance.

The `desktop` option was introduced before we had the `launched-from`
option. Now that `launched-from` is available, the `desktop` option is
no longer necessary (and sometimes harmful).

`ghostty +new-window` will still be slightly faster at getting a new
window since it does less initialization before it sends the signal to
the existing process but this will eliminate _a lot_ of confusion from
users.