gtk: why codegen when you can embed
When you do `glib-compile-resources` with `--generate-header` or
`--generate-source`, it generates really long and stinky C source files
littered with lengthy escape codes, macros and compiler directives, just
because C doesn't (yet!) have a standard way to just embed a file as a
literal bunch of bytes.

However, we do have access to that in Zig. So why not just do the same
thing but better?