lib-vt: wasm convenience functions and a simple example
This adds a set of Wasm convenience functions to ease memory management.
These are all prefixed with `ghostty_wasm` and are documented as part of
the standard Doxygen docs.

I also added a very simple single-page HTML example that demonstrates
how to use the Wasm module for key encoding. 

This also adds a bunch of safety checks to the C API to verify
that valid values are actually passed to the function. This is an
easy to hit bug.