terminal: HTML formatting
This adds HTML formatting capabilities to the formatter package. HTML
is emitted as inline styles. For palette indexes, direct RGB is emitted
if we have access to a palette; otherwise, we fall back to CSS
variables.

This isn't exposed to end users yet, but will enable copy as html
features. This is available in libghostty. 

Fixes #9395