perf: add custom spinlock to replace renderer_state.mutex
This eliminates a fairly large amount of overhead caused by the IO
thread having to go to sleep if it tries to lock the renderer state
while someone else (usually the renderer thread) has it. The renderer
thread, especially after recent changes, shouldn't have the lock held
for more than a few thousand cycles typically, which means that it's
incredibly wasteful to go to sleep instead of spinning.

This spinlock does a spin-then-block lock, where after it spins for a
while (10K tries in this impl) it gives up and blocks, deferring to a
futex instead. This means that if our guess is wrong and the lock is
being held for a longer time than we thought, it's not the end of the
world since the thread will yield to the kernel relatively fast in the
grand scheme of things. The amount of time spent spinning is roughly
equivalent to the overhead caused by blocking the thread, so worst case
scenario we double the lock time on an operation that is already longer
running, so the locking time itself is unlikely to be a big part of.

Benchmarks show big improvements with this change for me on macOS on an
Apple M3 Max CPU. I am hoping that these improvements also manifest on
other platforms and different CPUs, though behavior can be strange with
concurrent code like this, so we'll find out.